dispense replicas for job
