+++ pkg/util/helper/binding_test.go
@@ -24,7 +24,162 @@ import (
 	"github.com/karmada-io/karmada/pkg/util/fedinformer/keys"
 	"github.com/karmada-io/karmada/pkg/util/gclient"
 	"github.com/karmada-io/karmada/pkg/util/names"
+	testhelper "github.com/karmada-io/karmada/test/helper"
 )
+
+const (
+	ClusterMember1 = "member1"
+	ClusterMember2 = "member2"
+	ClusterMember3 = "member3"
+)
+
+func TestDispenseReplicasByTargetClusters(t *testing.T) {
+	type args struct {
+		clusters []workv1alpha2.TargetCluster
+		sum      int32
+	}
+	tests := []struct {
+		name string
+		args args
+		want []workv1alpha2.TargetCluster
+	}{
+		{
+			name: "empty clusters",
+			args: args{
+				clusters: []workv1alpha2.TargetCluster{},
+				sum:      10,
+			},
+			want: []workv1alpha2.TargetCluster{},
+		},
+		{
+			name: "1 cluster, 5 replicas, 10 sum",
+			args: args{
+				clusters: []workv1alpha2.TargetCluster{
+					{
+						Name:     ClusterMember1,
+						Replicas: 5,
+					},
+				},
+				sum: 10,
+			},
+			want: []workv1alpha2.TargetCluster{
+				{
+					Name:     ClusterMember1,
+					Replicas: 10,
+				},
+			},
+		},
+		{
+			name: "3 cluster, 1:1:1, 12 sum",
+			args: args{
+				clusters: []workv1alpha2.TargetCluster{
+					{
+						Name:     ClusterMember1,
+						Replicas: 5,
+					},
+					{
+						Name:     ClusterMember2,
+						Replicas: 5,
+					},
+					{
+						Name:     ClusterMember3,
+						Replicas: 5,
+					},
+				},
+				sum: 12,
+			},
+			want: []workv1alpha2.TargetCluster{
+				{
+					Name:     ClusterMember1,
+					Replicas: 4,
+				},
+				{
+					Name:     ClusterMember2,
+					Replicas: 4,
+				},
+				{
+					Name:     ClusterMember3,
+					Replicas: 4,
+				},
+			},
+		},
+		{
+			name: "3 cluster, 1:1:1, 10 sum",
+			args: args{
+				clusters: []workv1alpha2.TargetCluster{
+					{
+						Name:     ClusterMember1,
+						Replicas: 5,
+					},
+					{
+						Name:     ClusterMember2,
+						Replicas: 5,
+					},
+					{
+						Name:     ClusterMember3,
+						Replicas: 5,
+					},
+				},
+				sum: 10,
+			},
+			want: []workv1alpha2.TargetCluster{
+				{
+					Name:     ClusterMember1,
+					Replicas: 4,
+				},
+				{
+					Name:     ClusterMember2,
+					Replicas: 3,
+				},
+				{
+					Name:     ClusterMember3,
+					Replicas: 3,
+				},
+			},
+		},
+		{
+			name: "3 cluster, 1:2:3, 13 sum",
+			args: args{
+				clusters: []workv1alpha2.TargetCluster{
+					{
+						Name:     ClusterMember1,
+						Replicas: 1,
+					},
+					{
+						Name:     ClusterMember2,
+						Replicas: 2,
+					},
+					{
+						Name:     ClusterMember3,
+						Replicas: 3,
+					},
+				},
+				sum: 13,
+			},
+			want: []workv1alpha2.TargetCluster{
+				{
+					Name:     ClusterMember1,
+					Replicas: 2,
+				},
+				{
+					Name:     ClusterMember2,
+					Replicas: 4,
+				},
+				{
+					Name:     ClusterMember3,
+					Replicas: 7,
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := SpreadReplicasByTargetClusters(tt.args.sum, tt.args.clusters, nil); !testhelper.IsScheduleResultEqual(got, tt.want) {
+				t.Errorf("SpreadReplicasByTargetClusters() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
 
 func TestHasScheduledReplica(t *testing.T) {
 	tests := []struct {
